Implement dry command
One might to know what happens for a particular backup or restore. The
dry command let's you inspect this without any changes to the system.

+++ backup.sh
@@ -4,7 +4,7 @@
 
 exclude_file=.backup_ignore
 
-while getopts ":e:s:d:r:" option; do
+while getopts ":e:s:d:r:x" option; do
     case "${option}" in
         e)
             exclude_file=${OPTARG};;
@@ -14,6 +14,8 @@ while getopts ":e:s:d:r:" option; do
             source_paths=${OPTARG};;
         r)
             source_root=${OPTARG};;
+        x)
+            dry_run="--dry-run";;
         :)
             echo "Missing argument for option $OPTARG"
             exit 1;;
@@ -43,7 +45,7 @@ destination_root=$(trim_right_slash $destination_root)
 datetime="$(date '+%Y%m%d_%H%M%S')"
 
 backup_path="$destination_root/$datetime"
-mkdir -p $backup_path
+[ -z "$dry_run" ] && mkdir -p $backup_path
 latest_link="$destination_root/latest"
 
 [ ! -L $latest_link ] && echo "First backup! Link to latest previous backup does not exist, it will be created."
@@ -52,16 +54,18 @@ for source_path in $source_paths; do
     source_suffix=$(trim_right_slash "$(trim_left_slash "$source_path")")
     backup_path_suffix=$(trim_right_slash "$(trim_to_first_right_slash "$source_suffix")")
     destination="$backup_path/$backup_path_suffix"
-    mkdir -p $destination
+    [ -z "$dry_run" ] && mkdir -p $destination
     source="$source_root/$source_suffix"
 
-    rsync -aE --progress --delete \
+    rsync -aE --progress --delete $dry_run \
         --link-dest "$latest_link/$backup_path_suffix" \
         --exclude-from "$exclude_file" \
         "$source" \
         "$destination"
 done
 
-rm -rf $latest_link
-ln -s $backup_path $latest_link
+if [ -z "$dry_run" ]; then
+    rm -rf $latest_link
+    ln -s $backup_path $latest_link
+fi
